Add game module
This module will control the flow of the game. Several functions were
added to the module. The switchTurn function will switch the turns of
the players after a move has been made. The isWin function will check
if a player has won the game and the isTie function will check if there
has been a tie.